Bump version to 1.0.40 in config.json; update data type conversion in get_rms and detect_note_from_fft functions in listener.py

+++ shome_audio_listener/listener.py
@@ -31,7 +31,7 @@ for i in range(21, 109):  # MIDI 21–108
 
 # 📐 RMS hesaplama (sade)
 def get_rms(data):
-    samples = np.frombuffer(data, dtype=np.int16)
+    samples = np.frombuffer(data, dtype=np.int16).astype(np.float32)
     if samples.size == 0 or np.isnan(np.mean(samples)):
         return 0.0
     rms = np.sqrt(np.mean(samples**2))
@@ -39,7 +39,7 @@ def get_rms(data):
 
 # 🎼 FFT ile nota tahmini (opsiyonel)
 def detect_note_from_fft(data):
-    samples = np.frombuffer(data, dtype=np.int16)
+    samples = np.frombuffer(data, dtype=np.int16).astype(np.float32)
     if samples.size == 0:
         return None
